♻️ refactor: alterando caminho de diretorio do view
