+++ backend/src/main/java/com/app/guimscore/view/model/SignInReqDto.java
@@ -1,4 +1,4 @@
-package com.app.guimscore.controller.models;
+package com.app.guimscore.view.model;
 
 public record SignInReqDto(
         String name,
