+++ backend/src/main/java/com/app/guimscore/view/model/SignUpReqDto.java
@@ -1,11 +1,7 @@
-package com.app.guimscore.controller.models;
+package com.app.guimscore.view.model;
 
-import com.app.guimscore.model.GameServerModel;
-import com.app.guimscore.model.roles.UserRole;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotNull;
-
-import java.util.UUID;
 
 public record SignUpReqDto(
         @NotNull(message = "É necessario preencher o nome")
