+++ backend/src/main/java/com/app/guimscore/view/model/SignUpResDto.java
@@ -1,4 +1,4 @@
-package com.app.guimscore.controller.models;
+package com.app.guimscore.view.model;
 
 public record SignUpResDto(
         String message,
