+++ backend/src/test/java/com/app/guimscore/service/AuthServiceTest.java
@@ -1,0 +1,24 @@
+package com.app.guimscore.service;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+@DisplayName("test auth service")
+public class AuthServiceTest {
+
+    @DisplayName("test function signUp")
+    class SignUpTest {
+
+
+
+    }
+
+    @DisplayName("test function signIn")
+    class SignInTest {
+
+
+    }
+
+}
